requirements.txt
numpy==1.26.4
matplotlib==3.9.0
mne==1.7.0
customtkinter==5.2.2
scikit-learn==1.5.0
dearpygui==1.11.1
hydra-core==1.3.2
mne-lsl==1.4.0
memory-profiler==0.61.0
pylsl==1.16.2
pyriemann==0.6